Renamed mappings to maps
